Minor changes before ejecting
