+++ config-overrides.js
@@ -9,6 +9,9 @@ const options = {
 };
 
 module.exports = (config, env) => {
+  // Add decorator support
+  config = injectBabelPlugin(['transform-decorators-legacy'], config);
+
   if (process.env.ANALYZE) {
     // Run bundle analyzer
     config.plugins.push(new BundleAnalyzer());
@@ -22,9 +25,6 @@ module.exports = (config, env) => {
 
   // Remove data-testid properties
   config = injectBabelPlugin(['react-remove-properties', { properties: ['data-testid'] }], config);
-
-  // Add decorator support
-  config = injectBabelPlugin(['transform-decorators-legacy'], config);
 
   // Use custom Ant Design variables
   config = updateConfig(config, env, options);
