+++ package.json
@@ -42,6 +42,8 @@
     "typography": "^0.16.17"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-decorators": "^7.0.0-rc.1",
+    "@babel/preset-stage-2": "^7.0.0-rc.1",
     "babel-eslint": "^8.2.6",
     "babel-plugin-import": "^1.8.0",
     "babel-plugin-react-remove-properties": "^0.2.5",
@@ -62,6 +64,7 @@
     "prettier": "^1.14.2",
     "react-app-rewire-antd-theme": "^1.1.3",
     "react-app-rewire-less": "^2.1.2",
+    "react-app-rewire-mobx": "^1.0.8",
     "react-app-rewired": "^2.0.0",
     "react-styleguidist": "^7.2.1",
     "react-testing-library": "^4.1.9",
@@ -87,7 +90,7 @@
     "production": [
       ">0.25%",
       "not op_mini all",
-      "not ie 11"
+      "not ie > 0"
     ]
   }
 }
